requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
Django==4.1.1
django-crispy-forms==1.14.0
exceptiongroup==1.0.1
h11==0.14.0
idna==3.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-leetcode==1.2.1
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.2
urllib3==1.26.12
wsproto==1.2.0
